Ispravke nakon sastanka 22.02.2021.

+++ .eleventy.js
@@ -73,7 +73,7 @@ module.exports = function (config) {
         var dobj = new Date(dateObj);
         var day = dobj.getDate();
         var year = dobj.getFullYear();
-        var month = dobj.toLocaleString('hr-HR', { month: 'short' });
+        var month = dobj.toLocaleString('hr-HR', { month: 'long' });
         //var month = dobj.getMonth()+1;
         return day + ". " + month + " " + year + ".";
     });
